refactor: change import path to alias
