+++ src/api/Summary.ts
@@ -1,4 +1,4 @@
-﻿import { http } from "./Fetch";
+﻿import { http } from "@/api/Fetch";
 
 interface SummaryResponse {
   summaryId: number;
