+++ src/assets/index.ts
@@ -1,0 +1,28 @@
+﻿export { default as IconFAB } from "@/assets/icon/icon_fab.svg?react";
+export { default as IconFavorite } from "@/assets/icon/icon_favorite.svg?react";
+export { default as IconDownload } from "@/assets/icon/icon_download.svg?react";
+export { default as IconCheck } from "@/assets/icon/icon_check.svg?react";
+export { default as IconMenu } from "@/assets/icon/icon_menu.svg?react";
+export { default as IconQuestion } from "@/assets/icon/icon_question.svg?react";
+export { default as IconHome } from "@/assets/icon/icon_home.svg?react";
+export { default as IconTrash } from "@/assets/icon/icon_trash.svg?react";
+export { default as IconStore } from "@/assets/icon/icon_store.svg?react";
+export { default as IconAdd } from "@/assets/icon/icon_adddir.svg?react";
+export { default as IconUser} from "@/assets/icon/icon_user.svg?react";
+export { default as IconChevronDown } from "@/assets/icon/icon_chevron_down.svg?react";
+export { default as IconX } from "@/assets/icon/icon_x.svg?react";
+export { default as IconOauth} from "@/assets/icon/icon_google.svg?react";
+export { default as IconEdit } from "@/assets/icon/icon_edit.svg?react";
+export { default as IconArrowLeft } from "@/assets/icon/icon_arrow_left.svg?react";
+export { default as IconFilter } from "@/assets/icon/icon_filter.svg?react";
+export { default as IconGallery } from "@/assets/icon/icon_grid.svg?react";
+export { default as IconList } from "@/assets/icon/icon_list.svg?react";
+export { default as IconDeletion } from "@/assets/icon/icon_delete.svg?react";
+export { default as IconDocs } from "@/assets/icon/icon_docs.svg?react";
+export { default as IconNote } from "@/assets/icon/icon_note.svg?react";
+
+export { default as loadingAnimation } from "@/assets/animation/loading.json";
+
+export { default as LandingImg_1 } from "@/assets/image/image_landing_1.svg?react";
+export { default as LandingImg_2 } from "@/assets/image/image_landing_2.svg?react";
+export { default as LandingImg_3 } from "@/assets/image/image_landing_3.svg?react";
