+++ src/hooks/useTagOptions.ts
@@ -1,5 +1,5 @@
-﻿import { useTagStore } from "../store/useTagStore";
-import { type TagData } from "../types/tag";
+﻿import { useTagStore } from "@/store/useTagStore";
+import { type TagData } from "@/types/tag";
 
 export const useTagOptions = (): TagData[] => {
   const tags = useTagStore((state) => state.tags);
