+++ src/main.tsx
@@ -1,7 +1,7 @@
 import { Profiler, StrictMode } from "react";
 import { createRoot } from "react-dom/client";
-import "./styles/reset.css";
-import App from "./App.tsx";
+import "@/styles/reset.css";
+import App from "@/App.tsx";
 
 createRoot(document.getElementById("root")!).render(
   <StrictMode>
