+++ src/services/AppLayout.tsx
@@ -1,8 +1,8 @@
 ﻿import React, { useEffect } from "react";
 import { useNavigate } from "react-router-dom";
-import { useAuthStore } from "../store/useAuthStore";
-import Sidebar from "../commons/layout/sidebar/SideBar";
-import ScrollToTop from "../commons/layout/ScrollToTop";
+import { useAuthStore } from "@/store/useAuthStore";
+import Sidebar from "@/commons/layout/sidebar/SideBar";
+import ScrollToTop from "@/commons/layout/ScrollToTop";
 
 interface AppLayoutProps {
   children: React.ReactNode;
