+++ src/services/contents/note/NoteEditPage.tsx
@@ -1,21 +1,18 @@
 ﻿import { useState } from "react";
-import { useNavigate, useLocation, useParams } from "react-router-dom";
-import IconArrowLeft from "../../../assets/icon/icon_arrow_left.svg?react";
-import TextSection from "../../../commons/section/TextSection";
+import { useParams, useNavigate, useLocation } from "react-router-dom";
 import MDEditor from "@uiw/react-md-editor";
 import { getCodeString } from "rehype-rewrite";
 import katex from "katex";
-import "katex/dist/katex.min.css";
-import Button from "../../../commons/inputs/Button";
-import {
-  updateNoteTitle,
-  updateNoteTag,
-  updateNoteContent,
-} from "../../../api/Note";
-import { useTagStore } from "../../../store/useTagStore";
-import { useDirectoryStore } from "../../../store/useDirectoryStore";
-import type { TagData } from "../../../types/tag";
-import TagItem from "../../../commons/tag/TagItem";
+
+import TextSection from "@/commons/section/TextSection";
+import Button from "@/commons/inputs/Button";
+import TagItem from "@/commons/tag/TagItem";
+
+import { IconArrowLeft } from "@/assets";
+import type { TagData } from "@/types/tag";
+import { useTagStore } from "@/store/useTagStore";
+import { useDirectoryStore } from "@/store/useDirectoryStore";
+import { updateNoteContent, updateNoteTag, updateNoteTitle } from "@/api/Note";
 
 function NoteEditPage() {
   const navigate = useNavigate();
